move benchmark tests into a separate class
